feat(llm): full switch to vercel ai sdk
