+++ packages/llm/src/completion.builder.test.ts
@@ -1,231 +1,122 @@
-import assert from "node:assert/strict";
-import { describe, it } from "node:test";
+import { tool } from "ai";
+import { MockLanguageModelV1 } from "ai/test";
+import { describe } from "node:test";
 import { z } from "zod";
-import { assertExtends, assertType } from "../../shared/src/ts.utils.ts";
-import { AnthropicRunner } from "./anthropic/anthropic.runner.ts";
-import { CompletionBuilder } from "./completion.builder.ts";
-import { type CompletionRunner } from "./completion.runner.ts";
-import { type Llm } from "./llm.types.ts";
+import { assertType } from "../../shared/src/ts.utils.ts";
+import { completion } from "./completion.builder.ts";
+import type { Llm } from "./llm.types.ts";
 
 describe("CompletionBuilder", () => {
-  const _types = () => {
-    // It infers multi tools call structure
-    const multiToolCOmpletion = CompletionBuilder.new
-      .tools([
-        {
-          name: "test" as const,
-          schema: z.object({ value: z.string() }),
-        },
-        {
-          name: "test2" as const,
-          schema: z.object({ value2: z.string() }),
-        },
-      ])
-      .stopReason("tool_call");
-    const multiToolCalls = multiToolCOmpletion.lastMessageToolCalls();
+  const _typesTesting = async () => {
+    // Assert base completion
+    const builder = completion;
+    assertType<{}>(builder.$);
 
+    // Assert configured completion
+    const expected = {
+      maxTokens: 100,
+      temperature: 1,
+      topP: 2,
+      topK: 2,
+      frequencyPenalty: 3,
+      presencePenalty: 4,
+      seed: 5,
+      stopSequences: ["a", "b"],
+      maxRetries: 1,
+      abortSignal: new AbortSignal(),
+      maxSteps: 1,
+    } satisfies Llm.Completion.Partial;
+
+    const configuredBuilder = builder
+      .maxTokens(expected.maxTokens)
+      .temperature(expected.temperature)
+      .topP(expected.topP)
+      .topK(expected.topK)
+      .frequencyPenalty(expected.frequencyPenalty)
+      .presencePenalty(expected.presencePenalty)
+      .seed(expected.seed)
+      .stopSequences(expected.stopSequences)
+      .maxRetries(expected.maxRetries)
+      .abortSignal(expected.abortSignal)
+      .maxSteps(expected.maxSteps);
+
+    assertType<typeof expected>(configuredBuilder.$);
+
+    // Generation typings
+
+    // @ts-expect-error Cannot generate text without model and prompt
+    void configuredBuilder.generateText();
+    // @ts-expect-error Cannot stream text without model and prompt
+    void configuredBuilder.streamText();
+    // @ts-expect-error Cannot generate object without model and prompt
+    void configuredBuilder.generateObject();
+    // @ts-expect-error Cannot stream object without model and prompt
+    void configuredBuilder.streamObject();
+
+    const withMessages = configuredBuilder
+      .model(new MockLanguageModelV1({}))
+      .messages([]);
+
+    const object = await withMessages.generateObject({
+      schema: z.object({
+        a: z.string(),
+        b: z.number(),
+      }),
+    });
+    assertType<{ a: string; b: number }>(object.object);
+
+    const stream = withMessages.streamObject({
+      schema: z.object({
+        a: z.string(),
+        b: z.number(),
+      }),
+    });
+    assertType<Partial<{ a: string; b: number }>>(await stream.object);
+
+    const text = await withMessages.generateText();
+    assertType<string>(text.text);
+    assertType<Array<never>>(text.toolCalls);
+
+    const streamText = await withMessages.streamText();
+    assertType<string>(await streamText.text);
+    assertType<Array<never>>(await streamText.toolCalls);
+
+    // Tools typings
+
+    const withTools = withMessages.tools({
+      myTool: tool({
+        parameters: z.object({
+          a: z.string(),
+        }),
+      }),
+    });
+
+    // @ts-expect-error Tool is not in the tools
+    withTools.activeTools([""]);
+
+    withTools.activeTools(["myTool"]);
+
+    withTools.toolChoice({
+      // @ts-expect-error Tool is not in the tools
+      toolName: "plop",
+      type: "tool",
+    });
+
+    withTools.toolChoice({
+      toolName: "myTool",
+      type: "tool",
+    });
+
+    const textWithTools = await withTools.generateText();
     assertType<
-      Array<
-        | {
-            toolName: "test";
-            type: "tool_call";
-            toolCallId: string;
-            toolArgs: { value: string };
-          }
-        | {
-            toolName: "test2";
-            type: "tool_call";
-            toolCallId: string;
-            toolArgs: { value2: string };
-          }
-      >
-    >(multiToolCalls);
-
-    // It infers single tools call structure
-    const singleToolCOmpletion = CompletionBuilder.new
-      .tool({
-        name: "test" as const,
-        schema: z.object({ value: z.string() }),
-      })
-      .stopReason("tool_call");
-
-    const singleToolCall = singleToolCOmpletion.lastMessageToolCalls();
-
-    assertType<
-      [
-        {
-          toolName: "test";
-          type: "tool_call";
-          toolCallId: string;
-          toolArgs: { value: string };
-        },
-      ]
-    >(singleToolCall);
-
-    // It infers multi tool call with requireToolUse
-    const multiToolCOmpletionWithRequireToolUse = CompletionBuilder.new
-      .tools(
-        [
-          {
-            name: "test" as const,
-            schema: z.object({ value: z.string() }),
-          },
-        ],
-        true,
-      )
-      .stopReason("tool_call");
-    const multiToolCallsWithRequireToolUse =
-      multiToolCOmpletionWithRequireToolUse.lastMessageToolCalls();
-
-    assertType<
-      [
-        ...any[],
-        {
-          toolName: "test";
-          type: "tool_call";
-          toolCallId: string;
-          toolArgs: { value: string };
-        },
-      ]
-    >(multiToolCallsWithRequireToolUse);
+      Array<{
+        type: "tool-call";
+        toolCallId: string;
+        toolName: "myTool";
+        args: {
+          a: string;
+        };
+      }>
+    >(textWithTools.toolCalls);
   };
-
-  it("can be created", () => {
-    const completion = CompletionBuilder.new;
-    const res = completion.$;
-
-    type ExpectedType = {};
-    assertExtends<typeof res, ExpectedType>();
-    assertExtends<ExpectedType, typeof res>();
-    assert.deepEqual(res, {});
-
-    const _typeTest = () => {
-      const runner = {} as any as AnthropicRunner.Instance.Ready;
-      // @ts-expect-error anthropic conversion should not be available in this state
-      void completion.run(runner);
-    };
-  });
-
-  it("add informations", () => {
-    const userMessage = {
-      role: "user" as const,
-      content: [{ type: "text" as const, text: "Hello World" }],
-    };
-    const assistantMessage = {
-      role: "assistant" as const,
-      content: [{ type: "text" as const, text: "Hello World" }],
-    };
-
-    const testTool = {
-      name: "test",
-      schema: z.object({ name: z.string() }),
-    };
-
-    const completion = CompletionBuilder.new
-      .system("Hello World")
-      .temperature(8)
-      .maxTokens(4000)
-      .topK(10)
-      .topP(0.5)
-      .stopSequences(["\n"])
-      .addMessages([userMessage])
-      .addMessages([assistantMessage])
-      .tools([testTool], true);
-
-    const res = completion.$;
-
-    type ExpectedType = {
-      messages: Array<Llm.Message>;
-      system: string;
-      temperature: number;
-      maxTokens: number;
-      toolsConfig: Llm.Completion.ToolConfig.Multi;
-      topK: number;
-      topP: number;
-      stopSequences: string[];
-    };
-
-    assertType<ExpectedType>(res);
-
-    const _typeTest = () => {
-      const runner = {} as any as CompletionRunner<Llm.Completion>;
-      // Anthropic conversion should be available in this state
-      void completion.run(runner);
-    };
-
-    assert.deepEqual(res, {
-      system: "Hello World",
-      temperature: 8,
-      maxTokens: 4000,
-      toolsConfig: {
-        type: "multi",
-        tools: [testTool],
-        requireToolUse: true,
-      },
-      topK: 10,
-      topP: 0.5,
-      stopSequences: ["\n"],
-      messages: [userMessage, assistantMessage],
-    });
-  });
-
-  it("can be created from an existing completion", () => {
-    const completion = CompletionBuilder.from({
-      system: "Hello World",
-      temperature: 8,
-      maxTokens: 4000,
-      topK: 10,
-      topP: 0.5,
-      stopSequences: ["\n"],
-      messages: [
-        {
-          role: "user" as const,
-          content: [{ type: "text" as const, text: "Hello World" }],
-        },
-        {
-          role: "assistant" as const,
-          content: [{ type: "text" as const, text: "Hello World" }],
-        },
-      ],
-    });
-
-    const res = completion.$;
-
-    type ExpectedType = {
-      messages: Array<Llm.Message>;
-      system: string;
-      temperature: number;
-      maxTokens: number;
-      topK: number;
-      topP: number;
-      stopSequences: string[];
-    };
-
-    assertType<ExpectedType>(res);
-
-    const _typeTest = () => {
-      const runner = {} as any as CompletionRunner<Llm.Completion>;
-      // Anthropic conversion should be available in this state
-      void completion.run(runner);
-    };
-
-    assert.deepEqual(res, {
-      system: "Hello World",
-      temperature: 8,
-      maxTokens: 4000,
-      topK: 10,
-      topP: 0.5,
-      stopSequences: ["\n"],
-      messages: [
-        {
-          role: "user",
-          content: [{ type: "text", text: "Hello World" }],
-        },
-        {
-          role: "assistant",
-          content: [{ type: "text", text: "Hello World" }],
-        },
-      ],
-    });
-  });
 });
