+++ packages/llm/src/llm.index.ts
@@ -2,6 +2,7 @@ export { CompletionBuilder, completion } from "./completion.builder.ts";
 export { type Llm } from "./llm.types.ts";
 export {
   assistantMsg,
+  systemMsg,
   userMsg,
   type MessageTemplate,
 } from "./message.builder.ts";
